contrib: pulse_tools_common + dump-template-hex + BUILD_TESTS note
